Remove three ships in three months post
I realized that this post is a bit lazy. It's just a low-down of things
that I've created so far. Hmm, I'd prefer that the Projects tab will
contain some long-form writeup of something I've worked on. Or atleast
not a listicle. I'll be putting this in the drafts instead to serve as a
reference.

Maybe I'll write about geomancer and tiffany in-depth in the future. If
that day comes I'll put it back to Projects.

On another note, I decided to keep the "Experiments in Neural Style
Transfer" in projects because there still seems to be traffic on that
post. I don't want to just make it suddenly disappear. Ideally that
should be in the Notebooks instead. But it's OK, it's a good archive of
things I've worked on back in grad school.

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>